requirements.txt
scipy==1.7.1
numpy==1.21.2
matplotlib==3.2.2
opencv_python==4.5.3.56
torch==1.8.1
torchvision==0.9.1
tqdm==4.62.3
Pillow==8.3.1
